[Silver V] Title: 다리 놓기, Time: 16 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Silver/1010. 다리 놓기/다리 놓기.cc
@@ -31,7 +31,7 @@
 
 using namespace std;
 
-long long DP[31][31];
+int DP[31][31];
 
 void SetDP()
 {
